Add files via upload
The objective of this project was to create a Random Forest model capable of decanting fraud in Credit card transections to minimize the losses of the business.
In this data mose of the transactions are legitimate only 429 fraud out of 284807 transactions which is only 0.17% of all the transactions, Thus it's a very unbalanced data. We have to build a model which can handling this type of data, in this case Random Forest classifier is one of the best choice.