Add files via upload
The Decision Tree classifier has shown promising results with an accuracy of 92.47% on the training data. The confusion matrix further helps in understanding the specific performance metrics, providing a comprehensive evaluation of the model. Moving forward, validating the model on test data and performing further optimizations can help ensure that the model generalizes well and performs robustly in real-world scenarios.