Add files via upload
This Project report is based on the Oxford Parkinson's Disease Detection Dataset which consists of speech recordings from individuals with and without Parkinson's disease. The aim of the project is to gain insight on the data using EDA and train different ML models and compare their accuracy to find out which one model is most suitable for early detection of the disease using ML model.